requirements.txt
openai
PyMuPDF
python-docx
fuzzywuzzy
python-Levenshtein
requests
streamlit
python-dotenv
fastapi
uvicorn